move subchannel metric e2e test
