+++ balancer/pickfirst/metrics_test.go
@@ -35,7 +35,6 @@ import (
 	"google.golang.org/grpc/internal"
 	"google.golang.org/grpc/internal/stubserver"
 	"google.golang.org/grpc/internal/testutils"
-	"google.golang.org/grpc/internal/testutils/stats"
 	testgrpc "google.golang.org/grpc/interop/grpc_testing"
 	testpb "google.golang.org/grpc/interop/grpc_testing"
 	"google.golang.org/grpc/resolver"
@@ -60,213 +59,6 @@ func init() {
     	}
   	]
 	}`, pickfirst.Name)
-}
-
-// TestPickFirstMetrics verifies that a pick_first channel emits the correct
-// subchannel metrics on connect and disconnect.
-func (s) TestPickFirstMetrics(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), defaultTestTimeout)
-	defer cancel()
-
-	ss := &stubserver.StubServer{
-		EmptyCallF: func(context.Context, *testpb.Empty) (*testpb.Empty, error) {
-			return &testpb.Empty{}, nil
-		},
-	}
-	ss.StartServer()
-	defer ss.Stop()
-
-	sc := internal.ParseServiceConfig.(func(string) *serviceconfig.ParseResult)(pfConfig)
-
-	r := manual.NewBuilderWithScheme("whatever")
-	r.InitialState(resolver.State{
-		ServiceConfig: sc,
-		Addresses:     []resolver.Address{{Addr: ss.Address}}},
-	)
-
-	tmr := stats.NewTestMetricsRecorder()
-	cc, err := grpc.NewClient(r.Scheme()+":///", grpc.WithStatsHandler(tmr), grpc.WithTransportCredentials(insecure.NewCredentials()), grpc.WithResolvers(r))
-	if err != nil {
-		t.Fatalf("NewClient() failed with error: %v", err)
-	}
-	defer cc.Close()
-
-	tsc := testgrpc.NewTestServiceClient(cc)
-	if _, err := tsc.EmptyCall(ctx, &testpb.Empty{}); err != nil {
-		t.Fatalf("EmptyCall() failed: %v", err)
-	}
-
-	if got, _ := tmr.Metric("grpc.subchannel.connection_attempts_succeeded"); got != 1 {
-		t.Errorf("Unexpected data for metric %v, got: %v, want: %v", "grpc.subchannel.connection_attempts_succeeded", got, 1)
-	}
-	if got, _ := tmr.Metric("grpc.subchannel.connection_attempts_failed"); got != 0 {
-		t.Errorf("Unexpected data for metric %v, got: %v, want: %v", "grpc.subchannel.connection_attempts_failed", got, 0)
-	}
-	if got, _ := tmr.Metric("grpc.subchannel.disconnections"); got != 0 {
-		t.Errorf("Unexpected data for metric %v, got: %v, want: %v", "grpc.subchannel.disconnections", got, 0)
-	}
-	if got, _ := tmr.Metric("grpc.subchannel.open_connections"); got != 1 {
-		t.Errorf("Unexpected data for metric %v, got: %v, want: %v", "grpc.subchannel.open_connections", got, 1)
-	}
-
-	ss.Stop()
-	testutils.AwaitState(ctx, t, cc, connectivity.Idle)
-	if got, _ := tmr.Metric("grpc.subchannel.disconnections"); got != 1 {
-		t.Errorf("Unexpected data for metric %v, got: %v, want: %v", "grpc.subchannel.disconnections", got, 1)
-	}
-	if got, _ := tmr.Metric("grpc.subchannel.open_connections"); got != -1 {
-		t.Errorf("Unexpected data for metric %v, got: %v, want: %v", "grpc.subchannel.open_connections", got, -1)
-	}
-}
-
-// TestPickFirstMetricsFailure tests the connection attempts failed metric. It
-// configures a channel and scenario that causes a pick first connection attempt
-// to fail, and then expects that metric to emit.
-func (s) TestPickFirstMetricsFailure(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), defaultTestTimeout)
-	defer cancel()
-
-	sc := internal.ParseServiceConfig.(func(string) *serviceconfig.ParseResult)(pfConfig)
-
-	r := manual.NewBuilderWithScheme("whatever")
-	r.InitialState(resolver.State{
-		ServiceConfig: sc,
-		Addresses:     []resolver.Address{{Addr: "bad address"}}},
-	)
-	grpcTarget := r.Scheme() + ":///"
-	tmr := stats.NewTestMetricsRecorder()
-	cc, err := grpc.NewClient(grpcTarget, grpc.WithStatsHandler(tmr), grpc.WithTransportCredentials(insecure.NewCredentials()), grpc.WithResolvers(r))
-	if err != nil {
-		t.Fatalf("NewClient() failed with error: %v", err)
-	}
-	defer cc.Close()
-
-	tsc := testgrpc.NewTestServiceClient(cc)
-	if _, err := tsc.EmptyCall(ctx, &testpb.Empty{}); err == nil {
-		t.Fatalf("EmptyCall() passed when expected to fail")
-	}
-
-	if got, _ := tmr.Metric("grpc.subchannel.connection_attempts_succeeded"); got != 0 {
-		t.Errorf("Unexpected data for metric %v, got: %v, want: %v", "grpc.subchannel.connection_attempts_succeeded", got, 0)
-	}
-	if got, _ := tmr.Metric("grpc.subchannel.connection_attempts_failed"); got != 1 {
-		t.Errorf("Unexpected data for metric %v, got: %v, want: %v", "grpc.subchannel.connection_attempts_failed", got, 1)
-	}
-	if got, _ := tmr.Metric("grpc.subchannel.disconnections"); got != 0 {
-		t.Errorf("Unexpected data for metric %v, got: %v, want: %v", "grpc.subchannel.disconnections", got, 0)
-	}
-}
-
-// TestPickFirstMetricsE2E tests subchannel metrics emitted during a pickfirst
-// connection lifecycle: failure, success, and disconnection.
-func (s) TestPickFirstMetricsE2E(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), defaultTestTimeout)
-	defer cancel()
-
-	ss := &stubserver.StubServer{
-		EmptyCallF: func(context.Context, *testpb.Empty) (*testpb.Empty, error) {
-			return &testpb.Empty{}, nil
-		},
-	}
-	ss.StartServer()
-	defer ss.Stop()
-
-	sc := internal.ParseServiceConfig.(func(string) *serviceconfig.ParseResult)(pfConfig)
-	r := manual.NewBuilderWithScheme("whatever")
-	r.InitialState(resolver.State{
-		ServiceConfig: sc,
-		Addresses:     []resolver.Address{{Addr: "bad address"}}},
-	) // Will trigger connection failed.
-
-	grpcTarget := r.Scheme() + ":///"
-	reader := metric.NewManualReader()
-	provider := metric.NewMeterProvider(metric.WithReader(reader))
-	mo := opentelemetry.MetricsOptions{
-		MeterProvider: provider,
-		Metrics:       opentelemetry.DefaultMetrics().Add("grpc.subchannel.disconnections", "grpc.subchannel.connection_attempts_succeeded", "grpc.subchannel.connection_attempts_failed"),
-	}
-
-	cc, err := grpc.NewClient(grpcTarget, opentelemetry.DialOption(opentelemetry.Options{MetricsOptions: mo}), grpc.WithTransportCredentials(insecure.NewCredentials()), grpc.WithResolvers(r))
-	if err != nil {
-		t.Fatalf("NewClient() failed with error: %v", err)
-	}
-	defer cc.Close()
-
-	tsc := testgrpc.NewTestServiceClient(cc)
-	if _, err := tsc.EmptyCall(ctx, &testpb.Empty{}); err == nil {
-		t.Fatalf("EmptyCall() passed when expected to fail")
-	}
-
-	r.UpdateState(resolver.State{
-		ServiceConfig: sc,
-		Addresses:     []resolver.Address{{Addr: ss.Address}},
-	}) // Will trigger successful connection metric.
-	if _, err := tsc.EmptyCall(ctx, &testpb.Empty{}, grpc.WaitForReady(true)); err != nil {
-		t.Fatalf("EmptyCall() failed: %v", err)
-	}
-
-	// Stop the server, that should send signal to disconnect, which will
-	// eventually emit disconnection metric before ClientConn goes IDLE.
-	ss.Stop()
-	testutils.AwaitState(ctx, t, cc, connectivity.Idle)
-	wantMetrics := []metricdata.Metrics{
-		{
-			Name:        "grpc.subchannel.connection_attempts_succeeded",
-			Description: "EXPERIMENTAL. Number of successful connection attempts.",
-			Unit:        "{attempt}",
-			Data: metricdata.Sum[int64]{
-				DataPoints: []metricdata.DataPoint[int64]{
-					{
-						Attributes: attribute.NewSet(attribute.String("grpc.target", grpcTarget)),
-						Value:      1,
-					},
-				},
-				Temporality: metricdata.CumulativeTemporality,
-				IsMonotonic: true,
-			},
-		},
-		{
-			Name:        "grpc.subchannel.connection_attempts_failed",
-			Description: "EXPERIMENTAL. Number of failed connection attempts.",
-			Unit:        "{attempt}",
-			Data: metricdata.Sum[int64]{
-				DataPoints: []metricdata.DataPoint[int64]{
-					{
-						Attributes: attribute.NewSet(attribute.String("grpc.target", grpcTarget)),
-						Value:      1,
-					},
-				},
-				Temporality: metricdata.CumulativeTemporality,
-				IsMonotonic: true,
-			},
-		},
-		{
-			Name:        "grpc.subchannel.disconnections",
-			Description: "EXPERIMENTAL. Number of times the selected subchannel becomes disconnected.",
-			Unit:        "{disconnection}",
-			Data: metricdata.Sum[int64]{
-				DataPoints: []metricdata.DataPoint[int64]{
-					{
-						Attributes: attribute.NewSet(attribute.String("grpc.target", grpcTarget)),
-						Value:      1,
-					},
-				},
-				Temporality: metricdata.CumulativeTemporality,
-				IsMonotonic: true,
-			},
-		},
-	}
-
-	gotMetrics := metricsDataFromReader(ctx, reader)
-	for _, metric := range wantMetrics {
-		val, ok := gotMetrics[metric.Name]
-		if !ok {
-			t.Fatalf("Metric %v not present in recorded metrics", metric.Name)
-		}
-		if !metricdatatest.AssertEqual(t, metric, val, metricdatatest.IgnoreTimestamp(), metricdatatest.IgnoreExemplars()) {
-			t.Fatalf("Metrics data type not equal for metric: %v", metric.Name)
-		}
-	}
 }
 
 func metricsDataFromReader(ctx context.Context, reader *metric.ManualReader) map[string]metricdata.Metrics {
